Refactor: Moved NviCandidatePage and extracted navigation logic
